pyproject.toml
[tool.poetry]
name = "pyavd-cli"
version = "0.1.0"
description = "A collection of cli scripts to use pyavd"
authors = ["Arista Networks <eosplus-dev@arista.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pyavd = "^4.5.0"
ansible-core = "^2.16.3"


[tool.poetry.group.dev.dependencies]
black = "^24.1.1"
isort = "^5.13.2"
flake8 = "^7.0.0"
flake8-docstrings = "^1.7.0"
pre-commit = "^3.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 130
preview = true

[tool.isort]
profile = "black"
